Add main files for the project
It was a difficult start for me because I do everything like the people in the video and in StackOverflow, but my code still didn't want to start working (namely, my bot didn't want to work for a full day). I spent my entire day trying to solve this problem, and I did it!
What did I do for this? I tried to connect to the telegram.API without any proxy server. And it didn't work. I have tried too many things and modules, such as https-proxy-agent, etc. Well, the day is almost over.... and I decided to try a VPN and it finally worked!

Except this, I added npm API's and .eslintrc, .prettierrc and .env.
Well, the next day will be good without these problems :D